Clarify control flow on gRPC interceptor
This flows more logically and makes clear when errors are returned, when
the handler is run, and when the "CompleteJob" hook is done.

Generally, when a success state is reached, it's preferable to "return
nil" to indicate that no error is being returned, instead of "return
err" where "err" may be nil.