#454
Feature in Daemon to restrict access to only certain users , this feature is useful,when you are
in a test environment and dont want everyone to make requests to your service.
Since this was flag was introduced to restrict users while in testing mode, we dont want this configuration to be mistakenly set on mainnet.
Also this can be used in Blockchain mode disabled as well !!